ipq807x: add kernel 5.9
This adds a basic 5.9 kernel config.
Its based of the generic OpenWrt config as the 5.4 from the target was way too messy to get it in line with the generic one and drop out all of the debug etc features.

This is obviously missing a lot of drivers that require DTS changes and probably some that dont.

Signed-off-by: Robert Marko <robert.marko@sartura.hr>

config ipq