ipq807x: xiaomi-ax3600: correctly handle boot flag from uboot
Xiaomi uboot use various flag to handle dual partition. Tweak the
bootcount and platform script to handle them and correctly switch the
partition on sysupgrade.
flag_boot_rootfs does always contain the current partition loaded and
can be used to get the state of the rootfs used. Alternatively the
rootfs name (rootfs for partition 0 and rootfs_1 for partition 1) can be
used.

Signed-off-by: Ansuel Smith <ansuelsmth@gmail.com>