mvebu: rb5009: enabe QCA8081 PHY support for 6.1
The required DT nodes are already in the RB5009 DTS, so we only
need to enable the relevant kernel symbol to include QCA8081
support.

After OpenWrt commit 997acc7f, we need to enable the QCA808X_PHY
symbol instead of the old AT803X from which it was split off.

Signed-off-by: Stijn Segers <foss@volatilesystems.org>